+++ oss2/api.py
@@ -446,7 +446,7 @@ class Bucket(_Base):
         :param headers: HTTP头部
         :type headers: 可以是dict，建议是oss2.CaseInsensitiveDict
 
-        :return: :class:`HeadObjectResult <oss2.models.HeadObjectResults>`
+        :return: :class:`HeadObjectResult <oss2.models.HeadObjectResult>`
 
         :raises: 如果Bucket不存在或者Object不存在，则抛出 :class:`NotFound <oss2.exceptions.NotFound>`
         """
@@ -928,7 +928,7 @@ class _UrlMaker(object):
     def __call__(self, bucket_name, key):
         self.type = _determine_endpoint_type(self.netloc, self.is_cname, bucket_name)
 
-        key = urlquote(key)
+        key = urlquote(key, '')
 
         if self.type == _ENDPOINT_TYPE_CNAME:
             return '{0}://{1}/{2}'.format(self.scheme, self.netloc, key)