+++ tests/test_api_base.py
@@ -8,14 +8,7 @@ import sys
 from common import *
 
 
-class TestApiBase(unittest.TestCase):
-    def __init__(self, *args, **kwargs):
-        super(TestApiBase, self).__init__(*args, **kwargs)
-
-    def setUp(self):
-        self.bucket = oss2.Bucket(oss2.Auth(OSS_ID, OSS_SECRET), OSS_ENDPOINT, OSS_BUCKET)
-        self.bucket.create_bucket()
-
+class TestApiBase(OssTestCase):
     if OSS_CNAME:
         def test_cname_bucket(self):
             bucket = oss2.Bucket(oss2.Auth(OSS_ID, OSS_SECRET), OSS_CNAME, OSS_BUCKET, is_cname=True)